chore: move updateQueue type
