+++ web/test/mock/withStatefulUserData.tsx
@@ -1,7 +1,6 @@
 import { UpdateQueueContext } from "@/contexts/updateQueueContext";
 import * as useUserModule from "@/lib/data-hooks/useUserData";
-import { UpdateQueue } from "@/lib/types/types";
-import { UpdateQueueFactory } from "@/lib/UpdateQueue";
+import { UpdateQueue, UpdateQueueFactory } from "@/lib/UpdateQueue";
 import { useMockRoadmap } from "@/test/mock/mockUseRoadmap";
 import {
   StatefulDataContext,
